fix: any tpye 제거 후 type 지정 및 input type 정리
